bhg/M_project/fastapi_mini_project/pyproject.toml
[project]
name = "fastapi-mini-project"
version = "0.1.0"
description = ""
authors = [
    {name = " Yan02251",email = "xcvbnmz1238@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "psycopg (>=3.2.5,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.1.1,<4.0.0)",
]

[tool.poetry]
packages = [{include = "fastapi_mini_project", from = "src"}]


[tool.poetry.group.prod.dependencies]
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.38"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
pydantic = "^2.10.6"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
